New Post: Simpler documentation
Notes:  this post should be called simpler documentation process.

This is the rough draft of how to set up the node.js server that ran
tararoys.com from 2012-2014(Aka the present day.) It turns out that
recreating my original setup is a pain, so this post is going to
be repurposed to 'how to set up jekyll.' I am learning how to set
up jekyll right now.