start writing post simpler-documentation-with-a-jekyll-blog.markdown
This is the raw transcript of my shell session.  It contains every
command I used to set up the jekyll blog. It also contains commentary
in the form of echos.  The commentary is basically my thought process
and notes.  It sometimes contains links that I found particularly useful.

In addition to this documentation, I also saved a folder with every
website I found useful during the blog setup process.  I saved local
copies of the webite, and numbered the saved webisties in the order
in which i found them.  As I saved them, I made the background color of
teh sections of the site that I found particularly useful orange.

The folder is called GettingStartedWithJekyll. That is the other half
of this log.

I also put a jekyll- header
---
layout: post
title:  "Simple Documentation Setup"
date:   2014-03-30 18:32:40
categories: jekyll update
---

at the top of the log.  That was not part of the original log.