Console log of every command used to set up a jekyll blog
I use the echo command as a way to comment on commands and to figure
things out.  So nearly every command that I don't understand has an echoed
comment before or after it.  The echos also have links to the most useful
resources I found and code snippets that I am executing.