Explain webpage higlighting and begin linking research pages
I begin to link all of the webpages I saved into the tutorial using
absolute paths.   Note: it took me awhile to figure out that
jekyll will not render these pages if try to save them in the pages
file, but will render them if I save them in the root directory.