Commenting on setting up github and adding a new post
This section is a bit confused. As you can see from the log, I apparently
added a new post to the blog, and then decided I wanted to push the
blog to github without checking to see if I successfully added a new
blog. This is interesting because it makes figuring out what is going
on in the log a challenge.