Linked rendering console logs with liquid.
this is how to get the console logs displayed.  This was everything
I needed to know to make the raw console log.

And that is the last of the documentation I saved. I've needed to
google several things since then to be able to write this 'nicer'
tutorial, but I forgot to capture them.

95% of the things I'm doing by hand, and therefore having to bugfix
later, can be fixed by automating this somehow.

This is a perfect example of the frustration, friction, and general nastiness of this whole process of programming.  http://www.chris-granger.com/2014/03/27/toward-a-better-programming/

This is rediculous, and I will figure out how to make learning to program as easy as playing zelda ocarina of time.